New names & credits for OS X About box
The standard OS X Java launcher uses the name of the main class as the About box, so I renamed the main class to `HearthStatsOsx` so that the menu says 'About HearthStatsOsx' instead of 'About Main' or 'About MainOsx'. Also added a Credits.rtf file that gets displayed in the About box.